+++ src/components/ExpenseDashboardPage.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+const ExpenseDashboardPage = () => (
+  <div>
+    This is from my dashboard component.
+  </div>
+);
+
+export default ExpenseDashboardPage;
