+++ src/components/Header.js
@@ -1,0 +1,19 @@
+import React from 'react';
+import { NavLink } from 'react-router-dom';
+
+const Header = () => (
+  <header>
+    <h1>Expensify</h1>
+    <button>
+      <NavLink to="/" activeClassName="is-active" exact={true}>Home</NavLink> {/* NavLink activeClassName lets customize with css */}
+    </button>
+    <button>
+      <NavLink to="/create" activeClassName="is-active">Create Expense</NavLink>
+    </button>
+    <button>
+      <NavLink to="/help" activeClassName="is-active">Help</NavLink>
+    </button>
+  </header>
+);
+
+export default Header;
