+++ src/components/HelpPage.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+const HelpPage = () => (
+  <div>
+    This is my help page component.
+  </div>
+);
+
+export default HelpPage;
